requirements.txt
aiogram==2.25.1
boto3==1.26.0
cryptography==41.0.3
python-dotenv
botocore